Initiated the solution for converting a given number to words
The project is intended to create a simple program to generate a
list of possible combinations of words for a given set of numbers
from a given dictionary of words.

The program will be written with a bruteforce approach at first and
progressively be refined in each iterations until a viable solution
can be found, which can generate the combinations within 1000ms.