List of words for the given number is generated
Current method generate_words, generates a list of words from the
letter combinations and matches them with the dictionary to provide
a list of word combinations for the given phone number.

Currently correct combinations of the words are not generated and
some empty arrays are generate as well. Common word combinations for
the given phone number is currently unfulfilled. The program is
incomplete.