Added Gemfile and RSpec gem and updated get_word_combination() method
Gemfile has been created and rspec has been defined as an dependency
for the project. /spec folder has been created to hold the test files.
get_word_combination method has been defined as a public method to be
called for testing and other purpose from the object.

Previous bruteforce approach to the program did not require heavy
testing thus rspec was not included as a dependency early on. As the
code needs to be optimized, rspec has been added for testing.

The previous methods were called from run() method which has been edited
out in favour of one public method for the user/test to call upon.